+++ src/Utils/Constraints.py
@@ -1,3 +1,7 @@
+# --------------------------------------------------------------------------- #
+#                                  FUNCTIONS                                  #
+# --------------------------------------------------------------------------- #
+
 c_st_dependency_balance_region = {
     'constraint_class': 'c_st_dependency_balance_region',
     'constraint_parameters': {
@@ -49,76 +53,3 @@ c_mt_dependency_tot_transaction_amount_credit_score = {
 }
 
 # --------------------------------------------------------------------------- #
-#                                  FUNCTIONS                                  #
-# --------------------------------------------------------------------------- #
-
-# c_st_dependency_balance_region = {
-#     'constraint_class': 'FixedIncrements',
-#     'constraint_parameters': {
-#         'column_name': 'Balance',
-#         'increment_value': 2
-#             # *2,
-#             # *1.5,
-#             # *0.5,
-#             # *0.1,
-#     }
-# }
-            
-# c_st_dependency_balance_employment_type = {
-#     'constraint_class': 'FixedIncrements',
-#     'constraint_parameters': {
-#         'column_name': 'Balance',
-#         'increment_value': 1.5
-#             # *1.5,
-#             # *0.5,
-#     }
-# }
-
-# c_st_dependency_tot_transaction_amount_credit_score = {
-#     'constraint_class': 'FixedIncrements',
-#     'constraint_parameters': {
-#         'column_name': 'Credit_Score',
-#         'increment_value': 2
-#             # *2,
-#             # *0.1,
-#     }  
-# }
-
-# # --------------------------------------------------------------------------- #
-
-# c_mt_dependency_balance_region = {
-#     'constraint_class': 'FixedIncrements',
-#     'table_name': 'd1', # for multi table synthesizers
-#     'constraint_parameters': {
-#         'column_name': 'Balance',
-#         'increment_value': 2
-#             # *2,
-#             # *1.5,
-#             # *0.5,
-#             # *0.1,
-#     }
-# }
-            
-# c_mt_dependency_balance_employment_type = {
-#     'constraint_class': 'FixedIncrements',
-#     'table_name': 'd1', # for multi table synthesizers
-#     'constraint_parameters': {
-#         'column_name': 'Balance',
-#         'increment_value': 1.5
-#             # *1.5,
-#             # *0.5,
-#     }
-# }
-
-# c_mt_dependency_tot_transaction_amount_credit_score = {
-#     'constraint_class': 'FixedIncrements',
-#     'table_name': 'd1', # for multi table synthesizers
-#     'constraint_parameters': {
-#         'column_name': 'Credit_Score',
-#         'increment_value': 2
-#             # *2,
-#             # *0.1,
-#     }  
-# }
-
-# # --------------------------------------------------------------------------- #
