+++ src/Utils/Correlator.py
@@ -1,10 +1,4 @@
 # --------------------------------------------------------------------------- #
-#                                   IMPORT                                    #
-# --------------------------------------------------------------------------- #
-
-import random
-
-# --------------------------------------------------------------------------- #
 #                                  FUNCTIONS                                  #
 # --------------------------------------------------------------------------- #
 
@@ -16,8 +10,6 @@ def dependency_balance_employment_type(data):
             _['Balance'] *= 0.5
         elif _['Employment_Type'] == 'Libero Professionista':
             _['Balance'] *= 1.5
-        # elif _['Employment_Type'] == 'Libero Professionista':
-        #     _['Balance'] *= random.choice([0.2, 0.5, 1.5, 2, 2.5, 10])
 
 def dependency_balance_region(data):
     for _ in data:
@@ -40,22 +32,5 @@ def dependency_tot_transaction_amount_credit_score(data):
             _['Credit_Score'] *= 1
         elif _['Tot_Transaction_Amount'] <= 10000:
             _['Credit_Score'] *= 0.2
-
-# --------------------------------------------------------------------------- #
-#                                    OLDER                                    #
-# --------------------------------------------------------------------------- #
-
-def dependency_income_age_range(data):
-    for _ in data:
-        if _['Age Range'] == '65':
-            _['Income'] = _['Income'] * 2.5
-        elif _['Age Range'] == '55':
-            _['Income'] = _['Income'] * 2
-        elif _['Age Range'] == '45':
-            _['Income'] = _['Income'] * 1.5
-        elif _['Age Range'] == '35':
-            _['Income'] = _['Income'] * 1
-        elif _['Age Range'] == '25':
-            _['Income'] = _['Income'] * 0.5
             
 # --------------------------------------------------------------------------- #
