+++ src/models/mt_HMA.py
@@ -4,6 +4,9 @@
 
 from sdv.multi_table import HMASynthesizer
 
+# External Functions
+from utils.Constraints import *
+
 # --------------------------------------------------------------------------- #
 #                                    MODEL                                    #
 # --------------------------------------------------------------------------- #
@@ -11,16 +14,28 @@ from sdv.multi_table import HMASynthesizer
 def HMA(RD, metadata):
     # Create an empty dictionary
     datasets = {}
-
     # Add the DataFrame to the dictionary
     datasets['d1'] = RD
-
     synthesizer = HMASynthesizer(metadata)
     synthesizer.validate(datasets)
 
+    # Add Constraints
+    synthesizer.load_custom_constraint_classes(
+        filepath='utils/Constraints.py',
+        class_names=['c_mt_dependency_balance_region']
+    )
+    synthesizer.load_custom_constraint_classes(
+        filepath='utils/Constraints.py',
+        class_names=['c_mt_dependency_balance_employment_type']
+    )
+    synthesizer.load_custom_constraint_classes(
+        filepath='utils/Constraints.py',
+        class_names=['c_mt_dependency_tot_transaction_amount_credit_score']
+    )
+
     #Initialize the SDV model and fit it to the DataFrame:
     synthesizer.fit(datasets)
-
+    
     # Generate synthetic data using the SDV model:
     return synthesizer.sample(scale=1)
